move code to config utils
