+++ tkltest/util/config_options.py
@@ -150,6 +150,7 @@ __options_spec = {
             'is_cli_option': False,
             'type': str,
             'default_value': '',
+            'relative_fix_type': 'paths_list_file',
             'help_message': 'file containing paths to jar files that represent the library dependencies of app'
         },
         'config_file': {
@@ -179,6 +180,7 @@ __options_spec = {
             'is_cli_option': False,
             'type': list,
             'default_value': [],
+            'relative_fix_type': 'paths_list',
             'help_message': 'list of paths to application classes'
         },
         'java_jdk_home': {
